added code to customize the events color based on event types
