+++ client/src/pages/Profile.js
@@ -1,6 +1,6 @@
 import React, { useEffect } from "react";
 import { useParams } from "react-router-dom";
-import RequestsCalendar from "../components/RequestsCalendar"; //where the user will be able to see all requests in a calendar view
+import RequestsCalendar from "../components/calendar/RequestsCalendar"; //where the user will be able to see all requests in a calendar view
 
 const Profile = () => {
   const { username } = useParams();
